wip(api): Remove API. Moved API docstrings to their respective internal classes.
This commit breaks setup of lib cores.
