+++ py2hwsw/scripts/fsm_gen.py
@@ -4,15 +4,13 @@
 #
 # SPDX-License-Identifier: MIT
 
-from api_base import convert2internal
-
 
 def generate_fsm(core):
     """Generate verilog code with the fsm of this module.
     returns: Generated verilog code
     """
     if core.fsm is not None:
-        return convert2internal(core.fsm).verilog_code
+        return core.fsm.verilog_code
 
     return ""
