+++ py2hwsw/scripts/iob_conf.py
@@ -5,13 +5,35 @@
 from dataclasses import dataclass
 
 from iob_base import fail_with_msg, parse_short_notation_text
-from api_base import internal_api_class
 
 
-@internal_api_class("user_api.api", "iob_conf")
 @dataclass
 class iob_conf:
-    """Py2HWSW's internal implementation of 'iob_conf' API class."""
+    """
+    Class to represent a configuration option.
+
+    Attributes:
+        name (str): Configuration identifier name.
+        kind (str): Configuration type:
+                    M: Verilog macro; the IP final user may redefine this macro, affecting all instances of the core.
+                    C: Verilog derived or constant macro; the IP final user may not redefine this macro.
+                    P: Verilog parameter; the IP final user may redefine this parameter for each instance of the core.
+                    D: Verilog derived or constant parameter; the IP final user may not redefine this parameter.
+                    L: Verilog local parameter; the IP final user may not redefine this parameter.
+        value (str | int | bool): Configuration value.
+        min_value (str | int): Minimum value supported by the configuration option (NA if not applicable).
+        max_value (str | int): Maximum value supported by the configuration option (NA if not applicable).
+        descr (str): Description of the configuration option.
+        doc_only (bool): If enabled, configuration option will only appear in documentation. Not in the verilog code.
+    """
+
+    name: str = ""
+    kind: str = "P"
+    value: str | int | bool = ""
+    min_value: str | int = "NA"
+    max_value: str | int = "NA"
+    descr: str = "Default description"
+    doc_only: bool = False
 
     def validate_attributes(self):
         if not self.name:
@@ -34,11 +56,28 @@ class iob_conf:
 
 
 #
-# API methods
+# Other Py2HWSW interface methods
 #
 
 
-def conf_from_dict(conf_dict):
+def create_conf_from_dict(conf_dict):
+    """
+    Function to create iob_conf object from dictionary attributes.
+
+    Attributes:
+        conf_dict (dict): dictionary with values to initialize attributes of iob_conf object.
+            This dictionary supports the following keys corresponding to the iob_conf attributes:
+            - name           -> iob_conf.name
+            - kind           -> iob_conf.kind
+            - value          -> iob_conf.value
+            - min_value      -> iob_conf.min_value
+            - max_value      -> iob_conf.max_value
+            - descr          -> iob_conf.descr
+            - doc_only       -> iob_conf.doc_only
+
+    Returns:
+        iob_conf: iob_conf object
+    """
     return iob_conf(**conf_dict)
 
 
@@ -56,5 +95,19 @@ def conf_text2dict(conf_text: str) -> dict:
     return parse_short_notation_text(conf_text, conf_text_flags)
 
 
-def conf_from_text(conf_text: str) -> iob_conf:
-    return conf_from_dict(conf_text2dict(conf_text))
+def create_conf_from_text(conf_text):
+    """
+    Function to create iob_conf object from short notation text.
+
+    Attributes:
+        conf_text (str): Short notation text. Object attributes are specified using the following format:
+            name [-t kind] [-v value] [-m min_value] [-M max_value] [-doc]
+            [-d descr]
+            Example:
+                DATA_W -t P -v 32 -m NA -M NA
+                -d 'Data bus width'
+
+    Returns:
+        iob_conf: iob_conf object
+    """
+    return create_conf_from_dict(conf_text2dict(conf_text))
