+++ py2hwsw/scripts/iob_parameter.py
@@ -12,25 +12,46 @@ from iob_base import (
     assert_attributes,
     find_obj_in_list,
     parse_short_notation_text,
+    empty_list,
 )
 
-from api_base import internal_api_class
 
-
-@internal_api_class("user_api.draft_api", "iob_parameter")
 @dataclass
 class iob_parameter:
-    """Class to represent a IOb Parameter option."""
+    """
+    Class that represents a IOb parameter attribute.
+
+    Attributes:
+        name (str): Identifier name for the IOb Parameter option.
+        val (Any): Value of the IOb Parameter option.
+        descr (str): Description of the IOb Parameter option.
+    """
+
+    name: str = ""
+    val: object = ""
+    descr: str = "Default description"
 
     def validate_attributes(self):
         if not self.name:
             fail_with_msg("Every IOb Parameter must have a name!")
 
 
-@internal_api_class("user_api.draft_api", "iob_parameter_group")
 @dataclass
 class iob_parameter_group:
-    """Class to represent a Group of IOb Parameters."""
+    """
+    Class that represents a group of IOb Parameters.
+
+    Attributes:
+        name (str): Identifier name for the group of IOb Parameters.
+        descr (str): Description of the IOb Parameter group.
+        iob_parameters (list): List of IOb Parameter objects.
+        doc_clearpage (bool): If enabled, the documentation table for this group will be terminated by a TeX '\clearpage' command.
+    """
+
+    name: str = ""
+    descr: str = "Default description"
+    iob_parameters: list = empty_list()
+    doc_clearpage: bool = False
 
     def validate_attributes(self):
         if not self.name:
@@ -97,11 +118,36 @@ def create_iob_parameter_group(core, *args, **kwargs):
         raise
 
 
-def iob_parameter_from_dict(iob_parameter_dict):
+#
+# Other Py2HWSW interface methods
+#
+
+
+def create_iob_parameter_from_dict(iob_parameter_dict):
+    """
+    Function to create iob_iob_parameter object from dictionary attributes.
+
+    Attributes:
+        iob_parameter_dict (dict): dictionary with values to initialize attributes of iob_iob_parameter object.
+            This dictionary supports the following keys corresponding to the iob_parameter attributes:
+            - name -> iob_parameter.name
+            - val -> iob_parameter.val
+            - descr -> iob_parameter.descr
+
+    Returns:
+        iob_parameter: iob_parameter object
+    """
     return iob_parameter(**iob_parameter_dict)
 
 
 def iob_parameter_text2dict(iob_parameter_text):
+    """Convert iob_parameter short notation text to dictionary.
+    Atributes:
+        iob_parameter_text (str): Short notation text. See `create_iob_parameter_from_text` for format.
+
+    Returns:
+        dict: Dictionary with iob_parameter attributes.
+    """
     iob_parameter_flags = [
         "name",
         ["-v", {"dest": "val"}],
@@ -110,15 +156,42 @@ def iob_parameter_text2dict(iob_parameter_text):
     return parse_short_notation_text(iob_parameter_text, iob_parameter_flags)
 
 
-def iob_parameter_from_text(iob_parameter_text):
-    return iob_parameter_from_dict(iob_parameter_text2dict(iob_parameter_text))
+def create_iob_parameter_from_text(iob_parameter_text):
+    """
+    Function to create iob_parameter object from short notation text.
+
+    Attributes:
+        iob_parameter_text (str): Short notation text. Object attributes are specified using the following format:
+            [name] [-v value] [-d descr]
+            Example:
+                my_param -v 42 -d 'My parameter description'
+
+    Returns:
+        iob_parameter: iob_parameter object
+    """
+    return create_iob_parameter_from_dict(iob_parameter_text2dict(iob_parameter_text))
 
 
-def iob_parameter_group_from_dict(iob_parameter_group_dict):
+def create_iob_parameter_group_from_dict(iob_parameter_group_dict):
+    """
+    Function to create iob_parameter_group object from dictionary attributes.
+
+    Attributes:
+        iob_parameter_group_dict (dict): dictionary with values to initialize attributes of iob_parameter_group object.
+            This dictionary supports the following keys corresponding to the iob_parameter_group attributes:
+            - name -> iob_parameter_group.name
+            - descr -> iob_parameter_group.descr
+            - iob_parameters -> iob_parameter_group.iob_parameters
+            - doc_clearpage -> iob_parameter_group.doc_clearpage
+
+    Returns:
+        iob_parameter_group: iob_parameter_group object
+    """
     # Convert dictionary elements to objects
     kwargs = iob_parameter_group_dict.copy()
     kwargs["iob_parameters"] = [
-        iob_parameter_from_dict(i) for i in iob_parameter_group_dict["iob_parameters"]
+        create_iob_parameter_from_dict(i)
+        for i in iob_parameter_group_dict["iob_parameters"]
     ]
     return iob_parameter_group(**kwargs)
 
@@ -142,7 +215,17 @@ def iob_parameter_group_text2dict(iob_parameter_group_text):
     return iob_parameter_group_dict
 
 
-def iob_parameter_group_from_text(iob_parameter_group_text):
-    return iob_parameter_group_from_dict(
+def create_iob_parameter_group_from_text(iob_parameter_group_text):
+    """
+    Function to create iob_parameter_group object from short notation text.
+
+    Attributes:
+        iob_parameter_group_text (str): Short notation text. Object attributes are specified using the following format:
+            TODO
+
+    Returns:
+        iob_parameter_group: iob_parameter_group object
+    """
+    return create_iob_parameter_group_from_dict(
         iob_parameter_group_text2dict(iob_parameter_group_text)
     )
