+++ setup.py
@@ -42,7 +42,6 @@ setup(
     author_email="info@iobundle.com",
     license="MIT",
     packages=["py2hwsw"],
-    # py_modules=["api"], # TODO: Figure out how to add py2hwsw api as an importable module
     entry_points={"console_scripts": ["py2hwsw = py2hwsw.main:main"]},
     package_data={"py2hwsw": ["**/*", ".*"]},
     include_package_data=True,
